pyproject.toml
[tool.poetry]
name = "coursework-6"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
django-apscheduler = "^0.6.2"
redis = "^5.0.7"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
